repas-1.1.0 branch merged ont top of tag v2.0.0.

+++ action/action-dsl/src/test/java/com/powsybl/action/dsl/ConditionDslLoaderTest.java
@@ -184,13 +184,13 @@ public class ConditionDslLoaderTest {
 
         // Only line1 is overloaded
         line1.getTerminal1().setP(600.0f).setQ(300.0f); // i = 1019.2061
-        float current1 = line1.getTerminal1().getI();
+        double current1 = line1.getTerminal1().getI();
         line1.newCurrentLimits1().setPermanentLimit(current1 - 100).add();
         evalAndAssert(false, "allOverloaded(['NHV1_NHV2_1','NHV1_NHV2_2'])");
 
         // Both lines are overloaded
         line2.getTerminal1().setP(600.0f).setQ(300.0f); // i = 1019.2061
-        float current2 = line2.getTerminal1().getI();
+        double current2 = line2.getTerminal1().getI();
         line2.newCurrentLimits1().setPermanentLimit(current2 - 100).add();
         evalAndAssert(true, "allOverloaded(['NHV1_NHV2_1','NHV1_NHV2_2'])");
